lib: tool: Moved table printing functions from lib to tool and replaced printf by log macro in lib.
