+++ src/lib/include/dynamic-config.h
@@ -66,7 +66,6 @@ void sja1105_dyn_l2_lookup_cmd_unpack(void *buf, struct
                                       sja1105_dyn_l2_lookup_cmd *cmd);
 int sja1105_mgmt_route_get(struct sja1105_spi_setup*, struct sja1105_mgmt_entry*, int index);
 int sja1105_mgmt_route_set(struct sja1105_spi_setup*, struct sja1105_mgmt_entry*, int index);
-void sja1105_mgmt_entry_show(struct sja1105_mgmt_entry *entry);
 
 int sja1105_mac_config_get(struct sja1105_spi_setup *spi_setup,
                            struct sja1105_mac_config_entry *entry,
