+++ src/lib/static-config/tables/general-params.c
@@ -156,13 +156,3 @@ void sja1105_general_params_entry_fmt_show(
 	formatted_append(print_buf, len, fmt, "IGNORE2STF  0x%" PRIX64, entry->ignore2stf);
 	formatted_append(print_buf, len, fmt, "TPID2       0x%" PRIX64, entry->tpid2);
 }
-
-void sja1105_general_params_entry_show(struct sja1105_general_params_entry *entry)
-{
-	char print_buf[MAX_LINE_SIZE];
-	char *fmt = "%s\n";
-
-	memset(print_buf, 0, MAX_LINE_SIZE);
-	sja1105_general_params_entry_fmt_show(print_buf, MAX_LINE_SIZE, fmt, entry);
-	puts(print_buf);
-}
