+++ src/lib/static-config/tables/l2-lookup-params.c
@@ -112,14 +112,3 @@ void sja1105_l2_lookup_params_entry_fmt_show(
 	formatted_append(print_buf, len, fmt, "NO_ENF_HOSTPRT 0x%" PRIX64, entry->no_enf_hostprt);
 	formatted_append(print_buf, len, fmt, "NO_MGMT_LEARN  0x%" PRIX64, entry->no_mgmt_learn);
 }
-
-void sja1105_l2_lookup_params_entry_show(struct sja1105_l2_lookup_params_entry *entry)
-{
-	char print_buf[MAX_LINE_SIZE];
-	char *fmt = "%s\n";
-
-	memset(print_buf, 0, MAX_LINE_SIZE);
-	sja1105_l2_lookup_params_entry_fmt_show(print_buf, MAX_LINE_SIZE, fmt, entry);
-	puts(print_buf);
-}
-
