+++ src/lib/static-config/tables/mac-config.c
@@ -176,14 +176,3 @@ sja1105_mac_config_entry_fmt_show(
 	formatted_append(print_buf, len, fmt, "INGMIRRPCP 0x%" PRIX64, entry->ingmirrpcp);
 	formatted_append(print_buf, len, fmt, "INGMIRRDEI 0x%" PRIX64, entry->ingmirrdei);
 }
-
-void sja1105_mac_config_entry_show(struct sja1105_mac_config_entry *entry)
-{
-	char print_buf[MAX_LINE_SIZE];
-	char *fmt = "%s\n";
-
-	memset(print_buf, 0, MAX_LINE_SIZE);
-	sja1105_mac_config_entry_fmt_show(print_buf, MAX_LINE_SIZE, fmt, entry);
-	puts(print_buf);
-}
-
