+++ src/lib/static-config/tables/schedule-entry-points.c
@@ -67,13 +67,3 @@ void sja1105_schedule_entry_points_entry_fmt_show(
 	formatted_append(print_buf, len, fmt, "DELTA      0x%" PRIX64, entry->delta);
 	formatted_append(print_buf, len, fmt, "ADDRESS    0x%" PRIX64, entry->address);
 }
-
-void sja1105_schedule_entry_points_entry_show(struct sja1105_schedule_entry_points_entry *entry)
-{
-	char print_buf[MAX_LINE_SIZE];
-	char *fmt = "%s\n";
-
-	memset(print_buf, 0, MAX_LINE_SIZE);
-	sja1105_schedule_entry_points_entry_fmt_show(print_buf, MAX_LINE_SIZE, fmt, entry);
-	puts(print_buf);
-}
