+++ src/lib/static-config/tables/schedule-params.c
@@ -71,14 +71,3 @@ void sja1105_schedule_params_entry_fmt_show(
 	print_array(subscheind_buf, entry->subscheind, 8);
 	formatted_append(print_buf, len, fmt, "SUBSCHEIND %s", subscheind_buf);
 }
-
-void sja1105_schedule_params_entry_show(struct sja1105_schedule_params_entry *entry)
-{
-	char print_buf[MAX_LINE_SIZE];
-	char *fmt = "%s\n";
-
-	memset(print_buf, 0, MAX_LINE_SIZE);
-	sja1105_schedule_params_entry_fmt_show(print_buf, MAX_LINE_SIZE, fmt, entry);
-	puts(print_buf);
-}
-
