+++ src/lib/static-config/tables/vlan-lookup.c
@@ -74,14 +74,3 @@ void sja1105_vlan_lookup_entry_fmt_show(
 	formatted_append(print_buf, len, fmt, "TAG_PORT   0x%" PRIX64, entry->tag_port);
 	formatted_append(print_buf, len, fmt, "VLANID     0x%" PRIX64, entry->vlanid);
 }
-
-void sja1105_vlan_lookup_entry_show(struct sja1105_vlan_lookup_entry *entry)
-{
-	char print_buf[MAX_LINE_SIZE];
-	char *fmt = "%s\n";
-
-	memset(print_buf, 0, MAX_LINE_SIZE);
-	sja1105_vlan_lookup_entry_fmt_show(print_buf, MAX_LINE_SIZE, fmt, entry);
-	puts(print_buf);
-}
-
