+++ src/tool/internal.h
@@ -34,6 +34,10 @@
 #include <common.h>
 #include <lib/include/staging-area.h>
 #include <lib/include/spi.h>
+#include <lib/include/static-config.h>
+#include <lib/include/dynamic-config.h>
+#include <lib/include/port-control.h>
+#include <lib/include/gtable.h>
 
 struct general_config {
 	char *staging_area;
@@ -76,6 +80,45 @@ void  linewise_concat(char **buffers, int count);
 int   read_array(char *array_str, uint64_t *array_val, int max_count);
 int   reliable_uint64_from_string(uint64_t *to, char *from, char**);
 int   reliable_double_from_string(double *to, char *from, char**);
+
+
+#define TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(table)                           \
+	void sja1105_##table##_entry_show(struct sja1105_##table##_entry*);        \
+
+/* show table functions */
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(l2_forwarding_params);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(l2_forwarding);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(l2_policing);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(mac_config);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(schedule_entry_points_params);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(schedule_entry_points);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(schedule_params);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(schedule);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(vlan_lookup);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(xmii_params);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(sgmii);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(vl_forwarding_params);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(vl_forwarding);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(vl_policing);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(vl_lookup);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(avb_params);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(general_params);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(mac_config);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(l2_lookup);
+TOOL_DEFINE_HEADERS_FOR_CONFIG_TABLE(l2_lookup_params);
+
+void sja1105_table_header_show(struct sja1105_table_header *hdr);
+void sja1105_mgmt_entry_show(struct sja1105_mgmt_entry *entry);
+void sja1105_dyn_l2_lookup_cmd_show(struct sja1105_dyn_l2_lookup_cmd *cmd);
+void sja1105_cbs_show(struct sja1105_cbs *cbs);
+void sja1105_cgu_pll_control_show(struct sja1105_cgu_pll_control *pll_control,
+                                  uint64_t device_id);
+void sja1105_cgu_mii_control_show(struct sja1105_cgu_mii_control *mii_control);
+void sja1105_cgu_idiv_show(struct sja1105_cgu_idiv *idiv);
+void sja1105_cfg_pad_mii_tx_show(struct sja1105_cfg_pad_mii_tx *pad_mii_tx);
+int sja1105_static_config_hexdump(void *buf);
+void gtable_hexdump(void *table, int len);
+void gtable_bitdump(void *table, int len);
 
 #define SJA1105_NETCONF_ROOT "sja1105"
 #define SJA1105_NETCONF_NS   "http://nxp.com/ns/yang/tsn/sja1105"
